requirements.txt
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
Flask==2.3.3
Flask-Bootstrap==3.3.7.1
Werkzeug==2.3.7
Pillow==10.0.0
psutil==5.9.0
